refactor: replace per-folder index files with single components barrel export
