+++ src/App.jsx
@@ -1,7 +1,6 @@
 import { Routes, Route } from 'react-router'
 import './app.css'
-import Header from './components/header'
-import Footer from './components/footer'
+import { Header, Footer } from './components'
 import HomeView from './views/home'
 import DeckView from './views/deck'
 import SignUpView from './views/signup'
