--- src/components/button/index.js
@@ -1,1 +1,0 @@
-export { default } from './Button.jsx';
