--- src/components/dropdown/index.js
@@ -1,1 +1,0 @@
-export { default } from './Dropdown.jsx';
