--- src/components/footer/index.js
@@ -1,5 +1,0 @@
-export { default } from "./Footer.jsx";
-
-// Re-export everything from this folder here.
-// This allows importing from 'components/footer' instead of 'components/footer/footer.jsx'
-// Example: import Footer from './components/footer'
