--- src/components/header/index.js
@@ -1,5 +1,0 @@
-export { default } from './Header.jsx';
-
-// Re-export everything from this folder here.
-// This allows importing from 'components/header' instead of 'components/header/header.jsx'
-// Example: import { Header, Hamburger } from './components/header';
