+++ src/components/index.js
@@ -1,0 +1,12 @@
+// Samlar och exporterar alla komponenter från en plats så att de kan
+// importeras med { Footer, Header } from "../components" istället för
+// att behöva ange varje komponents exakta filsökväg.
+
+export { default as Button } from "./button/Button.jsx";
+export { default as Dropdown } from "./header/dropdown/Dropdown.jsx";
+export { default as Footer } from "./footer/Footer.jsx";
+export { default as Header } from "./header/Header.jsx";
+export { default as CategoryIcon } from "./icons/CategoryIcon.jsx";
+export { default as InfoIcon } from "./icons/InfoIcon.jsx";
+export { default as LoginModal } from "./loginModal/LoginModal.jsx";
+export { default as Wordmark } from "./wordmark/Wordmark.jsx";
