--- src/components/loginModal/index.js
@@ -1,1 +1,0 @@
-export {default} from './LoginModal.jsx'
