--- src/components/wordmark/index.js
@@ -1,1 +1,0 @@
-export { default } from './Wordmark.jsx';
