--- src/views/deck/completion/index.js
@@ -1,1 +1,0 @@
-export { default } from "./CompletionScreen.jsx";
