--- src/views/deck/infoModal/index.js
@@ -1,1 +1,0 @@
-export {default} from './InfoModal.jsx'
